Move media.py and partial_file.py to lib/
