+++ lib/partial_file.py
